+++ core-notifications-fe/src/core-notifications/resources/notifications/index.ts
@@ -1,13 +1,15 @@
-import NotificationsActiveOutlinedIcon from '@mui/icons-material/NotificationsActiveOutlined';
-import { NotificationTable } from './TableList'
-import { NotificationShow } from './show'
-import { ListGuesser, ShowGuesser } from 'react-admin';
-import { NotificationEdit } from './edit';
+import NotificationsActiveOutlinedIcon from '@mui/icons-material/NotificationsActiveOutlined'
+import { ShowGuesser } from 'react-admin'
+import { AdminTable } from './AdminTable'
+import { NotificationEdit } from './edit'
+import { UserList } from './UserList'
 
 export class Notifications {
   static Name = 'Notification'
   static Icon = NotificationsActiveOutlinedIcon
-  static List = NotificationTable
-  static Show = ShowGuesser // NotificationShow
+  
   static Edit = NotificationEdit // NotificationShow
+  static List = AdminTable
+  static Show = ShowGuesser // NotificationShow
+  static UserList = UserList
 }