+++ test-core-fe/scripts/data.ts
@@ -8,12 +8,19 @@ import { mockNotifications } from '@moonlight-labs/core-notifications-fe/mock'
 
 import { mockUsers } from '../src/data/mockUsers'
 import { mockCompanies } from '../src/data/mockCompanies'
+import { mock } from 'node:test'
 
 const users = mockUsers(10)
 const companies = mockCompanies(5)
 const webhooks = mockWebhooks()
 
-const notifications = mockNotifications(40, () => faker.helpers.arrayElement(users))
+// ensure current user has notifications
+
+let notifications = mockNotifications(1, () => users[0], 'Task')
+notifications = [...notifications, ...mockNotifications(1, () => users[0], 'Simple')]
+notifications = [...notifications, ...mockNotifications(1, () => users[0], 'Global')]
+
+notifications = [...notifications, ...mockNotifications(40, () => faker.helpers.arrayElement(users))]
 
 const versions = mockVersions(40).map((version: any) => {
   const user = faker.helpers.arrayElement(users)